fix: unify component parsing and restore codeowner attribution
