+++ metrics/extension-component-metrics-2026-02-27-summary.json
@@ -37,5 +37,168 @@
   ],
   "newComponents": [],
   "noReplacementComponents": 24,
-  "totalNoReplacementInstances": 671
+  "totalNoReplacementInstances": 671,
+  "codeOwnerStats": {
+    "@HowardBraham": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/accounts-engineers": {
+      "mmdsInstances": 72,
+      "deprecatedInstances": 374,
+      "totalInstances": 446,
+      "migrationPercentage": "16.14",
+      "filesCount": 66
+    },
+    "@MetaMask/confirmations": {
+      "mmdsInstances": 103,
+      "deprecatedInstances": 631,
+      "totalInstances": 734,
+      "migrationPercentage": "14.03",
+      "filesCount": 164
+    },
+    "@MetaMask/core-extension-ux": {
+      "mmdsInstances": 141,
+      "deprecatedInstances": 830,
+      "totalInstances": 971,
+      "migrationPercentage": "14.52",
+      "filesCount": 136
+    },
+    "@MetaMask/core-platform": {
+      "mmdsInstances": 1,
+      "deprecatedInstances": 403,
+      "totalInstances": 404,
+      "migrationPercentage": "0.25",
+      "filesCount": 54
+    },
+    "@MetaMask/design-system-engineers": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/extension-platform": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/extension-privacy-reviewers": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/extension-security-team": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/metamask-assets": {
+      "mmdsInstances": 2,
+      "deprecatedInstances": 214,
+      "totalInstances": 216,
+      "migrationPercentage": "0.93",
+      "filesCount": 43
+    },
+    "@MetaMask/metamask-earn": {
+      "mmdsInstances": 3,
+      "deprecatedInstances": 0,
+      "totalInstances": 3,
+      "migrationPercentage": "100.00",
+      "filesCount": 1
+    },
+    "@MetaMask/new-networks": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/notifications": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 57,
+      "totalInstances": 57,
+      "migrationPercentage": "0.00",
+      "filesCount": 18
+    },
+    "@MetaMask/perps": {
+      "mmdsInstances": 634,
+      "deprecatedInstances": 76,
+      "totalInstances": 710,
+      "migrationPercentage": "89.30",
+      "filesCount": 49
+    },
+    "@MetaMask/policy-reviewers": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/qa": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/ramp": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/swaps-engineers": {
+      "mmdsInstances": 31,
+      "deprecatedInstances": 220,
+      "totalInstances": 251,
+      "migrationPercentage": "12.35",
+      "filesCount": 48
+    },
+    "@MetaMask/transactions": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/wallet-integrations": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/web3auth": {
+      "mmdsInstances": 211,
+      "deprecatedInstances": 284,
+      "totalInstances": 495,
+      "migrationPercentage": "42.63",
+      "filesCount": 46
+    },
+    "@dbrans": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@unknown": {
+      "mmdsInstances": 254,
+      "deprecatedInstances": 1691,
+      "totalInstances": 1945,
+      "migrationPercentage": "13.06",
+      "filesCount": 250
+    }
+  }
 }
