--- metrics/extension-component-metrics-2026-02-28-summary.json
@@ -1,204 +1,0 @@
-{
-  "project": "extension",
-  "date": "2026-02-28",
-  "mmdsInstances": 1452,
-  "deprecatedInstances": 3838,
-  "mmdsUsageTotal": 1434,
-  "totalInstances": 5272,
-  "migrationPercentage": "27.20",
-  "componentsTracked": 15,
-  "mmdsComponentsAvailable": 25,
-  "mmdsComponentsList": [
-    "AvatarAccount",
-    "AvatarBase",
-    "AvatarFavicon",
-    "AvatarGroup",
-    "AvatarIcon",
-    "AvatarNetwork",
-    "AvatarToken",
-    "BadgeCount",
-    "BadgeIcon",
-    "BadgeNetwork",
-    "BadgeStatus",
-    "BadgeWrapper",
-    "Blockies",
-    "Box",
-    "Button",
-    "ButtonBase",
-    "ButtonHero",
-    "ButtonIcon",
-    "Checkbox",
-    "Icon",
-    "Input",
-    "Jazzicon",
-    "Maskicon",
-    "Text",
-    "TextButton"
-  ],
-  "newComponents": [],
-  "noReplacementComponents": 24,
-  "totalNoReplacementInstances": 671,
-  "codeOwnerStats": {
-    "@HowardBraham": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/accounts-engineers": {
-      "mmdsInstances": 72,
-      "deprecatedInstances": 374,
-      "totalInstances": 446,
-      "migrationPercentage": "16.14",
-      "filesCount": 66
-    },
-    "@MetaMask/confirmations": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 2,
-      "totalInstances": 2,
-      "migrationPercentage": "0.00",
-      "filesCount": 1
-    },
-    "@MetaMask/core-extension-ux": {
-      "mmdsInstances": 117,
-      "deprecatedInstances": 774,
-      "totalInstances": 891,
-      "migrationPercentage": "13.13",
-      "filesCount": 129
-    },
-    "@MetaMask/core-platform": {
-      "mmdsInstances": 1,
-      "deprecatedInstances": 406,
-      "totalInstances": 407,
-      "migrationPercentage": "0.25",
-      "filesCount": 56
-    },
-    "@MetaMask/design-system-engineers": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/extension-platform": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/extension-privacy-reviewers": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/extension-security-team": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/metamask-assets": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/metamask-earn": {
-      "mmdsInstances": 3,
-      "deprecatedInstances": 0,
-      "totalInstances": 3,
-      "migrationPercentage": "100.00",
-      "filesCount": 1
-    },
-    "@MetaMask/new-networks": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/notifications": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 57,
-      "totalInstances": 57,
-      "migrationPercentage": "0.00",
-      "filesCount": 18
-    },
-    "@MetaMask/perps": {
-      "mmdsInstances": 634,
-      "deprecatedInstances": 76,
-      "totalInstances": 710,
-      "migrationPercentage": "89.30",
-      "filesCount": 49
-    },
-    "@MetaMask/policy-reviewers": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/qa": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/ramp": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/swaps-engineers": {
-      "mmdsInstances": 30,
-      "deprecatedInstances": 162,
-      "totalInstances": 192,
-      "migrationPercentage": "15.63",
-      "filesCount": 36
-    },
-    "@MetaMask/transactions": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/wallet-integrations": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/web3auth": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@dbrans": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@unknown": {
-      "mmdsInstances": 595,
-      "deprecatedInstances": 2929,
-      "totalInstances": 3524,
-      "migrationPercentage": "16.88",
-      "filesCount": 519
-    }
-  }
-}
