+++ metrics/index.json
@@ -1,5 +1,5 @@
 {
-  "lastUpdated": "2026-03-03T01:58:13.265Z",
+  "lastUpdated": "2026-03-03T02:04:40.386Z",
   "projects": {
     "mobile": [
       {
@@ -105,10 +105,6 @@
       {
         "date": "2026-02-27",
         "file": "mobile-component-metrics-2026-02-27-data.json"
-      },
-      {
-        "date": "2026-02-28",
-        "file": "mobile-component-metrics-2026-02-28-data.json"
       }
     ],
     "extension": [
@@ -215,15 +211,11 @@
       {
         "date": "2026-02-27",
         "file": "extension-component-metrics-2026-02-27-data.json"
-      },
-      {
-        "date": "2026-02-28",
-        "file": "extension-component-metrics-2026-02-28-data.json"
       }
     ]
   },
   "latest": {
-    "mobile": "mobile-component-metrics-2026-02-28-data.json",
-    "extension": "extension-component-metrics-2026-02-28-data.json"
+    "mobile": "mobile-component-metrics-2026-02-27-data.json",
+    "extension": "extension-component-metrics-2026-02-27-data.json"
   }
 }
