+++ metrics/mobile-component-metrics-2026-02-27-summary.json
@@ -50,5 +50,168 @@
     "TextField"
   ],
   "noReplacementComponents": 49,
-  "totalNoReplacementInstances": 948
+  "totalNoReplacementInstances": 948,
+  "codeOwnerStats": {
+    "@MetaMask/accounts-engineers": {
+      "mmdsInstances": 38,
+      "deprecatedInstances": 159,
+      "totalInstances": 197,
+      "migrationPercentage": "19.29",
+      "filesCount": 38
+    },
+    "@MetaMask/card": {
+      "mmdsInstances": 266,
+      "deprecatedInstances": 152,
+      "totalInstances": 418,
+      "migrationPercentage": "63.64",
+      "filesCount": 39
+    },
+    "@MetaMask/confirmations": {
+      "mmdsInstances": 71,
+      "deprecatedInstances": 314,
+      "totalInstances": 385,
+      "migrationPercentage": "18.44",
+      "filesCount": 119
+    },
+    "@MetaMask/core-platform": {
+      "mmdsInstances": 1,
+      "deprecatedInstances": 100,
+      "totalInstances": 101,
+      "migrationPercentage": "0.99",
+      "filesCount": 30
+    },
+    "@MetaMask/delegation": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/design-system-engineers": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/metamask-assets": {
+      "mmdsInstances": 103,
+      "deprecatedInstances": 164,
+      "totalInstances": 267,
+      "migrationPercentage": "38.58",
+      "filesCount": 55
+    },
+    "@MetaMask/metamask-earn": {
+      "mmdsInstances": 10,
+      "deprecatedInstances": 287,
+      "totalInstances": 297,
+      "migrationPercentage": "3.37",
+      "filesCount": 66
+    },
+    "@MetaMask/mobile-admins": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/mobile-core-ux": {
+      "mmdsInstances": 35,
+      "deprecatedInstances": 305,
+      "totalInstances": 340,
+      "migrationPercentage": "10.29",
+      "filesCount": 61
+    },
+    "@MetaMask/mobile-platform": {
+      "mmdsInstances": 3,
+      "deprecatedInstances": 1,
+      "totalInstances": 4,
+      "migrationPercentage": "75.00",
+      "filesCount": 2
+    },
+    "@MetaMask/notifications": {
+      "mmdsInstances": 4,
+      "deprecatedInstances": 75,
+      "totalInstances": 79,
+      "migrationPercentage": "5.06",
+      "filesCount": 24
+    },
+    "@MetaMask/perps": {
+      "mmdsInstances": 104,
+      "deprecatedInstances": 702,
+      "totalInstances": 806,
+      "migrationPercentage": "12.90",
+      "filesCount": 108
+    },
+    "@MetaMask/predict": {
+      "mmdsInstances": 409,
+      "deprecatedInstances": 149,
+      "totalInstances": 558,
+      "migrationPercentage": "73.30",
+      "filesCount": 67
+    },
+    "@MetaMask/qa": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/ramp": {
+      "mmdsInstances": 72,
+      "deprecatedInstances": 693,
+      "totalInstances": 765,
+      "migrationPercentage": "9.41",
+      "filesCount": 100
+    },
+    "@MetaMask/rewards": {
+      "mmdsInstances": 451,
+      "deprecatedInstances": 56,
+      "totalInstances": 507,
+      "migrationPercentage": "88.95",
+      "filesCount": 62
+    },
+    "@MetaMask/supply-chain": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 0,
+      "totalInstances": 0,
+      "migrationPercentage": "0.00",
+      "filesCount": 0
+    },
+    "@MetaMask/swaps-engineers": {
+      "mmdsInstances": 28,
+      "deprecatedInstances": 122,
+      "totalInstances": 150,
+      "migrationPercentage": "18.67",
+      "filesCount": 35
+    },
+    "@MetaMask/transactions": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 4,
+      "totalInstances": 4,
+      "migrationPercentage": "0.00",
+      "filesCount": 1
+    },
+    "@MetaMask/wallet-integrations": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 33,
+      "totalInstances": 33,
+      "migrationPercentage": "0.00",
+      "filesCount": 7
+    },
+    "@MetaMask/web3auth": {
+      "mmdsInstances": 0,
+      "deprecatedInstances": 11,
+      "totalInstances": 11,
+      "migrationPercentage": "0.00",
+      "filesCount": 3
+    },
+    "@unknown": {
+      "mmdsInstances": 460,
+      "deprecatedInstances": 921,
+      "totalInstances": 1381,
+      "migrationPercentage": "33.31",
+      "filesCount": 236
+    }
+  }
 }
