--- metrics/mobile-component-metrics-2026-02-28-summary.json
@@ -1,214 +1,0 @@
-{
-  "project": "mobile",
-  "date": "2026-02-28",
-  "mmdsInstances": 2055,
-  "deprecatedInstances": 3275,
-  "mmdsUsageTotal": 887,
-  "totalInstances": 4162,
-  "migrationPercentage": "21.31",
-  "componentsTracked": 25,
-  "mmdsComponentsAvailable": 35,
-  "mmdsComponentsList": [
-    "AvatarAccount",
-    "AvatarBase",
-    "AvatarFavicon",
-    "AvatarGroup",
-    "AvatarIcon",
-    "AvatarNetwork",
-    "AvatarToken",
-    "BadgeCount",
-    "BadgeIcon",
-    "BadgeNetwork",
-    "BadgeStatus",
-    "BadgeWrapper",
-    "Blockies",
-    "BottomSheetFooter",
-    "BottomSheetOverlay",
-    "Box",
-    "Button",
-    "ButtonAnimated",
-    "ButtonBase",
-    "ButtonIcon",
-    "Card",
-    "Checkbox",
-    "HeaderBase",
-    "Icon",
-    "Input",
-    "Jazzicon",
-    "Label",
-    "Maskicon",
-    "RadioButton",
-    "Skeleton",
-    "Text",
-    "TextButton",
-    "TextField",
-    "TextOrChildren",
-    "Toast"
-  ],
-  "newComponents": [],
-  "noReplacementComponents": 49,
-  "totalNoReplacementInstances": 948,
-  "codeOwnerStats": {
-    "@MetaMask/accounts-engineers": {
-      "mmdsInstances": 38,
-      "deprecatedInstances": 155,
-      "totalInstances": 193,
-      "migrationPercentage": "19.69",
-      "filesCount": 38
-    },
-    "@MetaMask/card": {
-      "mmdsInstances": 266,
-      "deprecatedInstances": 152,
-      "totalInstances": 418,
-      "migrationPercentage": "63.64",
-      "filesCount": 39
-    },
-    "@MetaMask/confirmations": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/core-platform": {
-      "mmdsInstances": 1,
-      "deprecatedInstances": 100,
-      "totalInstances": 101,
-      "migrationPercentage": "0.99",
-      "filesCount": 30
-    },
-    "@MetaMask/delegation": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/design-system-engineers": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/metamask-assets": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/metamask-earn": {
-      "mmdsInstances": 10,
-      "deprecatedInstances": 186,
-      "totalInstances": 196,
-      "migrationPercentage": "5.10",
-      "filesCount": 41
-    },
-    "@MetaMask/mobile-admins": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/mobile-core-ux": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/mobile-platform": {
-      "mmdsInstances": 3,
-      "deprecatedInstances": 0,
-      "totalInstances": 3,
-      "migrationPercentage": "100.00",
-      "filesCount": 1
-    },
-    "@MetaMask/notifications": {
-      "mmdsInstances": 4,
-      "deprecatedInstances": 75,
-      "totalInstances": 79,
-      "migrationPercentage": "5.06",
-      "filesCount": 24
-    },
-    "@MetaMask/perps": {
-      "mmdsInstances": 104,
-      "deprecatedInstances": 702,
-      "totalInstances": 806,
-      "migrationPercentage": "12.90",
-      "filesCount": 108
-    },
-    "@MetaMask/predict": {
-      "mmdsInstances": 409,
-      "deprecatedInstances": 149,
-      "totalInstances": 558,
-      "migrationPercentage": "73.30",
-      "filesCount": 67
-    },
-    "@MetaMask/qa": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/ramp": {
-      "mmdsInstances": 72,
-      "deprecatedInstances": 693,
-      "totalInstances": 765,
-      "migrationPercentage": "9.41",
-      "filesCount": 100
-    },
-    "@MetaMask/rewards": {
-      "mmdsInstances": 451,
-      "deprecatedInstances": 56,
-      "totalInstances": 507,
-      "migrationPercentage": "88.95",
-      "filesCount": 62
-    },
-    "@MetaMask/supply-chain": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/swaps-engineers": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/transactions": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@MetaMask/wallet-integrations": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 33,
-      "totalInstances": 33,
-      "migrationPercentage": "0.00",
-      "filesCount": 7
-    },
-    "@MetaMask/web3auth": {
-      "mmdsInstances": 0,
-      "deprecatedInstances": 0,
-      "totalInstances": 0,
-      "migrationPercentage": "0.00",
-      "filesCount": 0
-    },
-    "@unknown": {
-      "mmdsInstances": 697,
-      "deprecatedInstances": 1947,
-      "totalInstances": 2644,
-      "migrationPercentage": "26.36",
-      "filesCount": 536
-    }
-  }
-}
