+++ metrics/timeline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-03-03T01:58:13.264Z",
+  "generatedAt": "2026-03-03T02:04:40.385Z",
   "mobile": {
     "dates": [
       "2025-08-29",
@@ -27,8 +27,7 @@
       "2026-02-07",
       "2026-02-14",
       "2026-02-21",
-      "2026-02-27",
-      "2026-02-28"
+      "2026-02-27"
     ],
     "migrationPercentage": [
       2.96,
@@ -56,7 +55,6 @@
       19.11,
       18.44,
       20.29,
-      21.31,
       21.31
     ],
     "mmdsInstances": [
@@ -85,7 +83,6 @@
       1704,
       1659,
       1891,
-      2055,
       2055
     ],
     "deprecatedInstances": [
@@ -114,7 +111,6 @@
       3150,
       3150,
       3210,
-      3275,
       3275
     ],
     "totalInstances": [
@@ -143,11 +139,9 @@
       3894,
       3862,
       4027,
-      4162,
       4162
     ],
     "componentsFullyMigrated": [
-      0,
       0,
       0,
       0,
@@ -201,8 +195,7 @@
       8,
       8,
       8,
-      8,
-      9
+      8
     ],
     "componentsNotStarted": [
       21,
@@ -230,11 +223,9 @@
       19,
       19,
       19,
-      19,
-      16
+      19
     ],
     "totalComponents": [
-      25,
       25,
       25,
       25,
@@ -288,7 +279,6 @@
       25,
       26,
       33,
-      35,
       35
     ],
     "mmdsComponentsList": [
@@ -1012,43 +1002,6 @@
         "TextField",
         "TextOrChildren",
         "Toast"
-      ],
-      [
-        "AvatarAccount",
-        "AvatarBase",
-        "AvatarFavicon",
-        "AvatarGroup",
-        "AvatarIcon",
-        "AvatarNetwork",
-        "AvatarToken",
-        "BadgeCount",
-        "BadgeIcon",
-        "BadgeNetwork",
-        "BadgeStatus",
-        "BadgeWrapper",
-        "Blockies",
-        "BottomSheetFooter",
-        "BottomSheetOverlay",
-        "Box",
-        "Button",
-        "ButtonAnimated",
-        "ButtonBase",
-        "ButtonIcon",
-        "Card",
-        "Checkbox",
-        "HeaderBase",
-        "Icon",
-        "Input",
-        "Jazzicon",
-        "Label",
-        "Maskicon",
-        "RadioButton",
-        "Skeleton",
-        "Text",
-        "TextButton",
-        "TextField",
-        "TextOrChildren",
-        "Toast"
       ]
     ],
     "newComponents": [
@@ -1090,16 +1043,15 @@
       [
         "RadioButton",
         "TextField"
-      ],
-      []
+      ]
     ],
     "latestChange": {
-      "migrationPercentageChange": "0.00",
-      "mmdsInstancesChange": 0,
-      "deprecatedInstancesChange": 0,
+      "migrationPercentageChange": "1.02",
+      "mmdsInstancesChange": 164,
+      "deprecatedInstancesChange": 65,
       "componentsFullyMigratedChange": 0,
-      "componentsInProgressChange": 1,
-      "mmdsComponentsAvailableChange": 0
+      "componentsInProgressChange": 0,
+      "mmdsComponentsAvailableChange": 2
     }
   },
   "extension": {
@@ -1129,8 +1081,7 @@
       "2026-02-07",
       "2026-02-14",
       "2026-02-21",
-      "2026-02-27",
-      "2026-02-28"
+      "2026-02-27"
     ],
     "migrationPercentage": [
       2.37,
@@ -1158,7 +1109,6 @@
       22.69,
       26.16,
       26.91,
-      27.2,
       27.2
     ],
     "mmdsInstances": [
@@ -1187,7 +1137,6 @@
       1162,
       1400,
       1436,
-      1452,
       1452
     ],
     "deprecatedInstances": [
@@ -1216,7 +1165,6 @@
       3898,
       3900,
       3852,
-      3838,
       3838
     ],
     "totalInstances": [
@@ -1245,11 +1193,9 @@
       5042,
       5282,
       5270,
-      5272,
       5272
     ],
     "componentsFullyMigrated": [
-      0,
       0,
       0,
       0,
@@ -1303,8 +1249,7 @@
       12,
       12,
       12,
-      12,
-      13
+      12
     ],
     "componentsNotStarted": [
       9,
@@ -1332,11 +1277,9 @@
       5,
       5,
       5,
-      5,
-      2
+      5
     ],
     "totalComponents": [
-      15,
       15,
       15,
       15,
@@ -1389,7 +1332,6 @@
       24,
       24,
       24,
-      25,
       25,
       25
     ],
@@ -2063,33 +2005,6 @@
         "Maskicon",
         "Text",
         "TextButton"
-      ],
-      [
-        "AvatarAccount",
-        "AvatarBase",
-        "AvatarFavicon",
-        "AvatarGroup",
-        "AvatarIcon",
-        "AvatarNetwork",
-        "AvatarToken",
-        "BadgeCount",
-        "BadgeIcon",
-        "BadgeNetwork",
-        "BadgeStatus",
-        "BadgeWrapper",
-        "Blockies",
-        "Box",
-        "Button",
-        "ButtonBase",
-        "ButtonHero",
-        "ButtonIcon",
-        "Checkbox",
-        "Icon",
-        "Input",
-        "Jazzicon",
-        "Maskicon",
-        "Text",
-        "TextButton"
       ]
     ],
     "newComponents": [
@@ -2122,23 +2037,22 @@
       [
         "Input"
       ],
-      [],
       []
     ],
     "latestChange": {
-      "migrationPercentageChange": "0.00",
-      "mmdsInstancesChange": 0,
-      "deprecatedInstancesChange": 0,
+      "migrationPercentageChange": "0.29",
+      "mmdsInstancesChange": 16,
+      "deprecatedInstancesChange": -14,
       "componentsFullyMigratedChange": 0,
-      "componentsInProgressChange": 1,
+      "componentsInProgressChange": 0,
       "mmdsComponentsAvailableChange": 0
     }
   },
   "summary": {
-    "totalWeeks": 27,
+    "totalWeeks": 26,
     "dateRange": {
       "start": "2025-08-29",
-      "end": "2026-02-28"
+      "end": "2026-02-27"
     }
   }
 }
